format python scripts, require segfile sizes
